(CODEMGMT-279) Specify versions for Her and Rspec.
Before this commit, if a user got a newer version of Her it could cause
Rspec tests to fail. With this commit, users are required to use the
version of Her that works with the module.

The commit also updated the Rspec tests to work with Rspec 3.x and
specifies this requirement in the gemspec.