Add ORM for individual item (User/Module/Release)
Before the find method only returned the response. With this commit
the find method now returns and object that exposes the data of the
request or nil if the request fails.

Also adds some tests for the find method. Many still fail due to
the where and PaginatedCollection classes being unfinished.