(CODEMGMT-299) Lazy access for missing information
Before this commit, all relations were filled out with an API request
when they were accessed. With this commit the API request will only be
sent if an unavailable piece of information is accessed. It also adds
lazy collections which will do the same thing over an array of lazy
relations.