(FORGE-360) Add `allow_md5` param to Release#verify method
Since calling any MD5 method can cause an immediate segfault on
FIPS-enabled systems, we need to provide a way to avoid that call
altogether. This change adds a param to the PuppetForge::V3::Release
class' #verify method that allows the caller to disable fallback to MD5
in cases where the SHA-256 checksum is not available.