(MAINT) Move deploy to distinct build stage and update api_key
Previously, every cell in the matrix was trying to deploy at the end,
this should change it to only do one deploy after every cell of the test
matrix passes.

Also, the api_key was previously encrypted for travis-ci.org whereas the
current integration is using travis-ci.com which requires a different
encrypted value.