(PE-27712) Allow faraday to go to version 0.14.0
Users discovered an issue where faraday 0.13.1 does not correctly
respect proxy settings unless they are set in the environment. This gem
sets them programmatically, and in 0.13.1, that is ignored. This caused
issues with proxies configured for r10k. This commit updates the max
allowed version of faraday to inclue 0.14.0, where this issue was fixed.
It also excludes the broken 0.13.1 version.