(PF-2437) Prepend 'Bearer' string to authorization
Authorized requests to the Forge require an authorization header of
"Bearer <api-key>". This changes the connection process so that if a
user has set `PuppetForge::Connection.authorization` to something that
looks like a Forge API key (a 64-character hex string), it automatically
prepends the "Bearer " string. This will not take effect if the user
has already prepended "Bearer " themselves, or if the authorization
value doesn't fit the expected key format.

Also updates the README with basic information on authorization.