Remove unused gettext-setup dependency
This patch removes the dependency to gettext-setup. My understanding is
that this was a preparation to support multiple locales/languages within
the whole puppet ecosystem. I also heard that this plan was cancelled.
Also this gem has no locales. On the other side pulling gettext-setup in
raises several issues. The puppet gem starts it's locale logic if the
gem is present and that raises some errors, for example
https://tickets.puppetlabs.com/browse/MODULES-6598

Another issue is that https://github.com/puppetlabs/gettext-setup-gem
isn't maintained at all. The dependencies aren't up2date, no response
to open PRs since 1.5 years. This is also blocking the current vox
pupuli modulesync run.