Move to Faraday 2.x
It does not attempt to support Faraday 1.x as well because dependency
wise it would be impossible to pull in the correct middleware. That is
why a hard switch was made.

This moves the minimum Ruby version to 2.6 since that's what Faraday 2
has as the minimum version.

It also pulls in faraday-follow_redirects, which was previously part of
Faraday itself.

The registration is now done without a lamba since it's allowed to pass
in a constant.

CI is also modified to test on all actual Ruby version that currently
match the required_ruby_version in the gemspec. That is, 2.6 - 3.1.