Add LRU cache for HTTP response caching
This commit adds a new class, PuppetForge::LruCache, that is used
by PuppetForge::V3::Base to cache HTTP get responses that are
returned in the PuppetForge::V3::Base#request method. This should
ensure that most HTTP requests are locally cached and should improve
performance and reduce load on the upstream servers. The maximum
cache size defaults to 30, however it is configurable using the
environment variable PUPPET_FORGE_MAX_CACHE_SIZE.